# src/main/java/com/example/bankcards/dto/CreateCardRequest.java
package com.example.bankcards.dto;

import jakarta.validation.constraints.Future;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;

import java.time.LocalDate;

public class CreateCardRequest {

    @NotBlank(message = "Номер карты не должен быть пустым")
    @Pattern(regexp = "\\d{16}", message = "Номер карты должен содержать 16 цифр")
    private String number;

    @NotBlank(message = "Имя владельца не должно быть пустым")
    private String owner;

    @Future(message = "Срок действия должен быть в будущем")
    private LocalDate expirationDate;

    public CreateCardRequest() {}

    public String getNumber() {
        return number;
    }

    public void setNumber(String number) {
        this.number = number;
    }

    public String getOwner() {
        return owner;
    }

    public void setOwner(String owner) {
        this.owner = owner;
    }

    public LocalDate getExpirationDate() {
        return expirationDate;
    }

    public void setExpirationDate(LocalDate expirationDate) {
        this.expirationDate = expirationDate;
    }
}